Make the `sounds::sound_markers` more public
This is a temporary fix to unbreak the tiles builds, but is almost
certainly the wrong approach for general use.
